requirements.txt
streamlit
langchain
langchain-openai
google-generativeai
python-dotenv
groq